fix draw_image
