+++ fltk/src/draw.rs
@@ -787,10 +787,20 @@ pub unsafe fn draw_rgb_nocopy<T: WidgetBase>(wid: &mut T, fb: &[u8]) {
 
 /// Draw an image into a widget
 /// Requires a call to app::set_visual(Mode::Rgb).unwrap()
-/// # Safety
-/// The image should be a valid RGB image
-pub unsafe fn draw_image(data: &[u8], x: i32, y: i32, w: i32, h: i32, depth: i32) {
-    Fl_draw_image(data.as_ptr(), x, y, w, h, depth, 0)
+pub fn draw_image(data: &[u8], x: i32, y: i32, w: i32, h: i32, depth: u32) -> Result<(), FltkError> {
+    if depth > 4 {
+        return Err(FltkError::Internal(FltkErrorKind::ImageFormatError));
+    }
+    let mut sz = (w * h) as usize;
+    if depth > 0 {
+        sz *= depth as usize;
+    }
+    if sz > data.len() {
+        return Err(FltkError::Internal(FltkErrorKind::ImageFormatError));
+    }
+    unsafe {
+        Ok(Fl_draw_image(data.as_ptr(), x, y, w, h, depth as i32, 0))
+    }
 }
 
 /// Transforms raw data to png file
